Optimize digit counting in challenge 322
Optimized the method use to count the number of digits in an integer for
challenge 322 - Largest Palindrome. The previous method converted the
integer to a string and then found the length of the string; the new
method performs mathematical operations on the integer to determine the
number of digits it contains which improves the speed.