Fixed elevator Car movement bug in challenge 218
The logic behind elevator Car movement was causing unpredictable
behavior which resulted in an inconsistent time of completion. The logic
was adjusted such that all Riders are serviced in a consistent number of
"seconds"; the implemented logic may not necessarily be the most
advanced or fastest but it finishes in a consistent manner.

Fixed a bug which allowed a Car to Pickup more than its Capacity would
allow.

Cleaner and more idiomatic implementation of concurrency.